Fixed hover using v-btn instead of v-card
The problem is that the hover-event is somehow shared between all nested v-cards. I could not find a fix for this. This issue always occurs when nesting a component inside itself, so a solution would be to use another component (as we need v-card as our background), which I did here. The appearance is almost identical to the original approach.