Added basic Submission logic
Created a new file to cache the old submission page.
Exercises now have a button for submitting a solution. This button redirects to a page, where the create and submit a solution functionality will soon be added. Currently only testing basic functionality. Also waiting for the backend to implement submission types.