Mainly fixed the submissions
I adjusted the submissions to work with the new backend requests and data structures. Submissions should work as they did before.
In addition, I removed the ID-field in the exercise creator and added a dropdown to choose the type of exercise. However, the type "free-text-field" is the only type currently supported. The exercise editor does not have the type-dropdown to avoid issues (trying to switch the type of the exercise when submissions of the current type exist; a check could be implemented for this though).
I also added some currently unused functionality, which I therefore commented out, as I was having issues with the implementation (time-problem, as well as some breaking changes from the backend).